chore: shared showcase
