+++ apps/_components/src/index.ts
@@ -7,3 +7,4 @@ export { Figma } from './logos/Figma';
 export { Github } from './logos/Github';
 export { Slack } from './logos/Slack';
 export { Footer } from './Footer/Footer';
+export { Showcase } from './Showcase/Showcase';
