+++ packages/react/package.json
@@ -50,6 +50,7 @@
     "clsx": "^2.1.1"
   },
   "devDependencies": {
+    "@repo/components": "workspace:^",
     "@rollup/plugin-commonjs": "^26.0.1",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@testing-library/dom": "^10.4.0",
