+++ packages/react/stories/showcase.mdx
@@ -4,4 +4,4 @@ import * as ShowcaseStories from './showcase.stories';
 
 <Meta of={ShowcaseStories} />
 
-<Story of={ShowcaseStories.Showcase} />
+<Story of={ShowcaseStories.ShowcaseStory} />
