+++ packages/react/stories/showcase.module.css
@@ -1,14 +1,3 @@
-.components {
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: var(--ds-spacing-8);
-  max-width: 1440px;
-  margin: 0 auto;
-  background-color: var(--ds-color-neutral-background-default);
-  border-radius: var(--ds-border-radius-md);
-}
-
 .controls {
   display: grid;
   grid-template-columns: repeat(3, auto);
@@ -21,158 +10,4 @@
 .controls code {
   background: var(--ds-color-neutral-background-subtle);
   padding: var(--ds-spacing-1);
-}
-
-.card {
-  background-color: var(--ds-color-neutral-background-subtle);
-  border-radius: var(--ds-border-radius-md);
-  padding: 24px;
-}
-
-.cardTitle {
-  margin-bottom: 16px;
-}
-
-.checkbox {
-  grid-column: span 2;
-  padding-bottom: 16px;
-}
-
-.user {
-  grid-column: span 2;
-}
-
-.tableContainer {
-  grid-column: span 4;
-}
-
-.help {
-  grid-column: span 4;
-}
-
-.radio {
-  grid-column: span 2;
-}
-
-.tag {
-  grid-column: span 2;
-}
-
-.switches {
-  grid-column: span 2;
-}
-
-.combobox {
-  grid-column: span 2;
-}
-
-.tabs {
-  grid-column: span 4;
-  display: flex;
-  gap: 16px;
-  flex-direction: column;
-}
-
-.faq {
-  grid-column: span 4;
-}
-
-.userLink {
-  margin-top: 20px;
-}
-
-.userTitle {
-  text-align: center;
-}
-
-.userField {
-  margin-top: 24px;
-}
-
-.userBtn {
-  margin-top: 32px;
-}
-
-.tabs button {
-  min-height: 40px;
-}
-
-.table {
-  width: 100%;
-  margin-bottom: 24px;
-  margin-top: 16px;
-}
-
-.table thead button {
-  padding-top: 8px;
-  padding-bottom: 8px;
-}
-
-.tableCell {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: 13px 12px;
-}
-
-.tableImg {
-  height: 36px;
-  width: 36px;
-  border-radius: 50%;
-  box-shadow: 0 0 4px #353535;
-}
-
-.tableAction {
-  display: flex;
-  align-items: center;
-}
-
-.tableSelect {
-  width: 160px;
-}
-
-.tableHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 16px;
-}
-
-.tableSearch {
-  width: 260px;
-}
-
-.tableBtn {
-  margin-left: 20px;
-}
-
-.helpCards {
-  grid-column: 1 / 6;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 20px;
-}
-
-.helpHeading {
-  margin-bottom: 20px;
-}
-
-.tagList {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 16px;
-  margin-top: 16px;
-}
-
-.switchGroup {
-  display: flex;
-  gap: 26px;
-  flex-direction: column;
-}
-
-.toggleCombo {
-  display: flex;
-  flex-direction: column;
-  gap: var(--ds-spacing-2);
 }
